# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Junction Optimization System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  uptime_seconds?: number;
}

export interface LaneData {
  lane_id: number;
  vehicle_count: number;
  density: number;
  cars?: number;
  bikes?: number;
  trucks?: number;
  buses?: number;
  avg_speed?: number;
  congestion_level: string;
  direction?: string;
}

export type LiveTrafficStateLanes = { [key: string]: LaneData };

export interface LiveTrafficState {
  timestamp: string;
  demo_mode: boolean;
  lanes: LiveTrafficStateLanes;
}

export interface TrafficLog {
  id: number;
  lane_id: number;
  timestamp: string;
  vehicle_count: number;
  density: number;
  cars?: number;
  bikes?: number;
  trucks?: number;
  buses?: number;
  /** @nullable */
  avg_speed?: number | null;
  congestion_level: string;
}

export interface TrafficHistoryResponse {
  items: TrafficLog[];
  total: number;
}

export interface TrafficSummary {
  total_vehicles_today: number;
  avg_density: number;
  peak_hour: number;
  critical_events_today: number;
  emergency_events_today: number;
  system_uptime_seconds: number;
}

export type SignalStatePhases = { [key: string]: string };

export interface SignalState {
  phases: SignalStatePhases;
  /** @nullable */
  current_green: number | null;
  emergency_active: boolean;
  /** @nullable */
  emergency_lane: number | null;
  manual_override: boolean;
  time_remaining: number;
  ai_mode: boolean;
}

export interface SignalOverrideInput {
  lane_id: number;
  duration_seconds: number;
  emergency?: boolean;
  reason?: string;
}

export interface SignalOverrideResult {
  status: string;
  lane_id: number;
  /** @nullable */
  reason?: string | null;
}

export interface AiModeInput {
  enabled: boolean;
}

export interface AiModeResult {
  ai_mode: boolean;
}

export interface EmergencyEvent {
  id: number;
  junction_id?: number;
  lane_id: number;
  timestamp: string;
  vehicle_type: string;
  confidence: number;
  /** @nullable */
  duration_seconds?: number | null;
  resolved: boolean;
  /** @nullable */
  resolved_at?: string | null;
}

export interface EmergencyEventList {
  events: EmergencyEvent[];
  total: number;
}

export interface LanePrediction {
  lane_id: number;
  current_density: number;
  prediction_5min: number;
  prediction_10min: number;
  prediction_15min: number;
  alert_level: string;
}

export interface PredictionResponse {
  predictions: LanePrediction[];
  generated_at: string;
}

export interface SignalLogEntry {
  id: number;
  lane_id: number;
  timestamp: string;
  phase: string;
  green_time_sec: number;
  trigger: string;
  density_at_time: number;
}

export interface SignalLogSummary {
  ai_count: number;
  manual_count: number;
  emergency_count: number;
  avg_green_sec: number;
}

export interface SignalLogResponse {
  logs: SignalLogEntry[];
  total: number;
  summary: SignalLogSummary;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface LoginResult {
  access_token: string;
  token_type: string;
}

export type GetTrafficHistoryParams = {
  /**
   * @nullable
   */
  lane_id?: number | null;
  limit?: number;
};

export type GetSignalLogParams = {
  limit?: number;
  /**
   * @nullable
   */
  lane_id?: number | null;
};

export type GetEmergencyEventsParams = {
  limit?: number;
};