# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Junction Optimization System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signals/override:
    post:
      operationId: overrideSignal
      tags: [signals]
      summary: Override signal for a specific lane (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalOverrideInput"
      responses:
        "200":
          description: Override applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalOverrideResult"
  /signals/log:
    get:
      operationId: getSignalLog
      tags: [signals]
      summary: Get signal timing log with trigger breakdown and summary stats
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: lane_id
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Signal timing log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalLogResponse"
  /signals/ai-mode:
    post:
      operationId: setAiMode
      tags: [signals]
      summary: Toggle AI auto-management mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiModeInput"
      responses:
        "200":
          description: Mode updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModeResult"
...
components:
  schemas:
...
    SignalOverrideInput:
      type: object
      properties:
        lane_id:
          type: integer
        duration_seconds:
          type: integer
        emergency:
          type: boolean
        reason:
          type: string
      required:
        - lane_id
        - duration_seconds
    SignalOverrideResult:
      type: object
      properties:
        status:
          type: string
        lane_id:
          type: integer
        reason:
          type: ["string", "null"]
      required:
        - status
        - lane_id
    AiModeInput:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
    AiModeResult:
      type: object
      properties:
        ai_mode:
          type: boolean
      required:
        - ai_mode
...
    SignalLogEntry:
      type: object
      properties:
        id:
          type: integer
        lane_id:
          type: integer
        timestamp:
          type: string
        phase:
          type: string
        green_time_sec:
          type: integer
        trigger:
          type: string
        density_at_time:
          type: number
      required:
        - id
        - lane_id
        - timestamp
        - phase
        - green_time_sec
        - trigger
        - density_at_time
    SignalLogSummary:
      type: object
      properties:
        ai_count:
          type: integer
        manual_count:
          type: integer
        emergency_count:
          type: integer
        avg_green_sec:
          type: integer
      required:
        - ai_count
        - manual_count
        - emergency_count
        - avg_green_sec
    SignalLogResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/SignalLogEntry"
        total:
          type: integer
        summary:
          $ref: "#/components/schemas/SignalLogSummary"
      required:
        - logs
        - total
        - summary
...